Updates and tests for rebase.py
