+++ .travis.yml
@@ -1,0 +1,5 @@
+language: python
+python:
+    - 2.7
+
+script: ./run-tests.sh
