+++ testdata/5_input
@@ -1,0 +1,11 @@
+pick 5hfos3s [SQUASH][BUILD] Check in new shiny script for OSE
+pick c21ea24 [CARRY][BRANDING] OAuth server branding
+pick ffd6cab [SQUASH][BUILD] More diff for OSE tools
+pick 975a829 [CARRY][BUILD_GEN] Specfile updates
+pick 59dn2sf [CARRY][BUILD] Tooling diff for OSE
+pick 22a76cd [DROP] bump origin-web-console 0c5b53c
+pick ba426b9 Automatic commit of package [atomic-openshift] release [3.5.0.50-1].
+pick 4365d5d [DROP][INVESTIGATE] Weird shell diffs
+pick 8ff2968 [DROP][FIXME] Change image build and drop this cluster up flag
+pick c21ea22 [SQUASH][BRANDING] Additional OAuth server branding
+pick testtha [SQUASH][BUILD] Even more tooling diff for OSE
